added a confirmation before deleting profiles

+++ scripts/home.js
@@ -15,9 +15,30 @@ $(async function () {
     });
 
     $(document).on("click", "#delete-profile-button", function () {
-        profileManager.deleteProfile($('#profileDropdown').text());
-        uiManager.loadProfileMenuData();
-        loadActionbars();
+        bootbox.confirm({
+            title: "Delete Profile",
+            message: "Are you sure you want to delete this profile?",
+            buttons: {
+            confirm: {
+                label: 'Yes',
+                className: 'btn-danger'
+            },
+            cancel: {
+                label: 'No',
+                className: 'btn-secondary'
+            }
+            },
+            callback: function (result) {
+            if (result) {
+                profileManager.deleteProfile($('#profileDropdown').text());
+                uiManager.loadProfileMenuData();
+                loadActionbars();
+                $("#export-button").addClass('d-none');
+                $("#delete-profile-button").addClass('d-none');
+                $("#add-actionbar-button").addClass('d-none');
+            }
+            }
+        });
     });
 
     $(document).on("click", "#set-endpoint-button", function () {
@@ -38,7 +59,136 @@ $(async function () {
         await loadActionbars();
         $("#export-button").removeClass('d-none');
         $("#delete-profile-button").removeClass('d-none');
-        $("#add-actionbar-button").removeClass('d-none');        
+        $("#add-actionbar-button").removeClass('d-none');
+    });
+
+    $(document).on("click", ".CompoundItem", function (e) {
+
+        let $compoundSlot = $(e.target).closest('.CompoundItem');
+        let $compoundSlotChildrenContainer = $compoundSlot.next('.actionbar-slot-details');
+        let elementDetailsDisplayed = $compoundSlotChildrenContainer.css('display') !== 'none';
+    
+        // If the slot details already exist and are hidden, slide them down
+        if ($compoundSlotChildrenContainer.length > 0 && !elementDetailsDisplayed) {
+            $compoundSlotChildrenContainer.slideDown();
+            return;
+        }
+    
+        // If the slot details already exist and are shown, toggle them with slideUp
+        if ($compoundSlotChildrenContainer.length > 0 && elementDetailsDisplayed) {
+            $compoundSlotChildrenContainer.slideUp();
+            return;
+        }
+    
+        let profileName = $('#profileDropdown').text();
+        let actionbarIndex = $compoundSlot.attr('data-actionbar-index');
+        let slotIndex = $compoundSlot.attr('data-slot-index');
+    
+        // Append a spinner while loading
+        $compoundSlot.find('.keybind').after('<div class="spinner-border text-primary" role="status" style="width:15px; height:15px;"><span class="visually-hidden">Loading...</span></div>');
+    
+        // Placeholder HTML structure for slots (before data is loaded)
+        let actionbarSlot = profileManager.getActionbarSlot(profileName, actionbarIndex, slotIndex);
+        let actionbarSlotHtml = '';
+        for (let actionIndex = 0; actionIndex < actionbarSlot.actions.length; actionIndex++) {
+            let placeholderColumn = htmlTemplateProvider.getSlotLoadingTemplate(actionbarIndex, actionIndex); // Placeholder template
+            actionbarSlotHtml += placeholderColumn;
+        }
+    
+        // Append placeholders to DOM
+        $compoundSlot.after(`<div class="actionbar-slot-details">${actionbarSlotHtml}</div>`);
+    
+        // Remove the spinner after adding placeholders
+        $compoundSlot.find('.spinner-border').remove();
+    
+        let promises = [];
+        for (let actionIndex = 0; actionIndex < actionbarSlot.actions.length; actionIndex++) {
+            promises.push(
+                profileManager.getActionbarSlotWithApiDataAndImageLink(profileName, actionbarIndex, slotIndex, actionIndex).then((actionWithImage) => {
+                    let actionHtml = htmlTemplateProvider.getChildSlotTemplate(actionbarIndex, slotIndex, actionWithImage);
+                    $(`.loading-slot[data-actionbar-index="${actionbarIndex}"][data-slot-index="${actionIndex}"]`).replaceWith(actionHtml);
+                })
+            );
+        }
+    
+        // Once all promises are resolved, show the content with a slide-down animation
+        Promise.all(promises).then(() => {
+            $compoundSlot.next('.actionbar-slot-details').slideDown();
+            let sortableContainer = $compoundSlot.next('.actionbar-slot-details')[0];
+            sortableManager.initializeSortableContainer(sortableContainer, actionbarIndex, slotIndex);
+        });
+    });    
+
+    $(document).on("click", "#export-button", function () {
+        const profileName = $('#profileDropdown').text();
+        let json = profileManager.exportProfile(profileName);
+        bootbox.alert(`
+            <h5 class="mb-4">To import this into the Actionbars plugin you can either shift & right click an existing actionbar slot or go got the Actionbar plugin settings > infoboxes section > import setup.</h5>
+            <textarea class="form-control" rows="20">${json}</textarea>`);
+    });
+
+    $(document).on("click", ".search-result-item", function () {
+
+        const actionbarIndex = $(this).attr('data-actionbar-index');
+        const slotIndex = $(this).attr('data-slot-index');
+        const actionIndex = $(this).attr('data-action-index');
+        const itemId = $(this).attr('data-item-id');
+        const imageLink = $(this).find('.search-image').attr('src');
+
+        itemFetcher.fetchItemActionsById(itemId).then((actions) => {
+
+            if (actions.length == 1) {
+
+                profileManager.updateSlotAction(actionbarIndex, slotIndex, actions[0], actionIndex)
+                uiManager.setSlotAction(actionbarIndex, slotIndex, actionIndex, actions[0]);
+                uiManager.setSlotImage(actionbarIndex, slotIndex, imageLink);
+
+            } else {
+
+                modalProvider.showActionSelectionModal(actions, actionbarIndex, slotIndex, itemId, actionIndex, imageLink);
+
+            }
+        });
+    });
+
+    $(document).on("click", ".add-new-slot", function () {
+
+        var actionbarIndex = $(this).attr('data-actionbar-index');
+        var slotIndex = $(this).attr('data-slot-index');
+
+        modalProvider.showAddNewSlotModal(actionbarIndex, slotIndex);
+    });
+
+    $(document).on("click", "#add-actionbar-button", function () {   
+        var profileName = $('#profileDropdown').text();
+        profileManager.addEmptyActionbar(profileName);
+        loadActionbars();
+    });
+
+    $(document).on("click", ".delete-actionbar", function () {
+        var actionbarIndex = $(this).attr('data-actionbar-index');
+        var profileName = $('#profileDropdown').text();
+        
+        bootbox.confirm({
+            title: `Delete actionbar ${parseInt(actionbarIndex) + 1}`,
+            message: "Are you sure you want to delete this actionbar?",
+            buttons: {
+                confirm: {
+                    label: 'Yes',
+                    className: 'btn-danger'
+                },
+                cancel: {
+                    label: 'No',
+                    className: 'btn-secondary'
+                }
+            },
+            callback: function (result) {
+                if (result) {
+                    profileManager.deleteActionbar(profileName, actionbarIndex);
+                    loadActionbars();
+                }
+            }
+        });
     });
 
     $(document).on("contextmenu", ".PrayerItem", function (e) {
@@ -193,135 +343,6 @@ $(async function () {
             contextMenuProvider.showContextMenu(e, menuOptions);
         });
 
-    });
-
-    $(document).on("click", ".CompoundItem", function (e) {
-
-        let $compoundSlot = $(e.target).closest('.CompoundItem');
-        let $compoundSlotChildrenContainer = $compoundSlot.next('.actionbar-slot-details');
-        let elementDetailsDisplayed = $compoundSlotChildrenContainer.css('display') !== 'none';
-    
-        // If the slot details already exist and are hidden, slide them down
-        if ($compoundSlotChildrenContainer.length > 0 && !elementDetailsDisplayed) {
-            $compoundSlotChildrenContainer.slideDown();
-            return;
-        }
-    
-        // If the slot details already exist and are shown, toggle them with slideUp
-        if ($compoundSlotChildrenContainer.length > 0 && elementDetailsDisplayed) {
-            $compoundSlotChildrenContainer.slideUp();
-            return;
-        }
-    
-        let profileName = $('#profileDropdown').text();
-        let actionbarIndex = $compoundSlot.attr('data-actionbar-index');
-        let slotIndex = $compoundSlot.attr('data-slot-index');
-    
-        // Append a spinner while loading
-        $compoundSlot.find('.keybind').after('<div class="spinner-border text-primary" role="status" style="width:15px; height:15px;"><span class="visually-hidden">Loading...</span></div>');
-    
-        // Placeholder HTML structure for slots (before data is loaded)
-        let actionbarSlot = profileManager.getActionbarSlot(profileName, actionbarIndex, slotIndex);
-        let actionbarSlotHtml = '';
-        for (let actionIndex = 0; actionIndex < actionbarSlot.actions.length; actionIndex++) {
-            let placeholderColumn = htmlTemplateProvider.getSlotLoadingTemplate(actionbarIndex, actionIndex); // Placeholder template
-            actionbarSlotHtml += placeholderColumn;
-        }
-    
-        // Append placeholders to DOM
-        $compoundSlot.after(`<div class="actionbar-slot-details">${actionbarSlotHtml}</div>`);
-    
-        // Remove the spinner after adding placeholders
-        $compoundSlot.find('.spinner-border').remove();
-    
-        let promises = [];
-        for (let actionIndex = 0; actionIndex < actionbarSlot.actions.length; actionIndex++) {
-            promises.push(
-                profileManager.getActionbarSlotWithApiDataAndImageLink(profileName, actionbarIndex, slotIndex, actionIndex).then((actionWithImage) => {
-                    let actionHtml = htmlTemplateProvider.getChildSlotTemplate(actionbarIndex, slotIndex, actionWithImage);
-                    $(`.loading-slot[data-actionbar-index="${actionbarIndex}"][data-slot-index="${actionIndex}"]`).replaceWith(actionHtml);
-                })
-            );
-        }
-    
-        // Once all promises are resolved, show the content with a slide-down animation
-        Promise.all(promises).then(() => {
-            $compoundSlot.next('.actionbar-slot-details').slideDown();
-            let sortableContainer = $compoundSlot.next('.actionbar-slot-details')[0];
-            sortableManager.initializeSortableContainer(sortableContainer, actionbarIndex, slotIndex);
-        });
-    });    
-
-    $(document).on("click", "#export-button", function () {
-        const profileName = $('#profileDropdown').text();
-        let json = profileManager.exportProfile(profileName);
-        bootbox.alert(`
-            <h5 class="mb-4">To import this into the Actionbars plugin you can either shift & right click an existing actionbar slot or go got the Actionbar plugin settings > infoboxes section > import setup.</h5>
-            <textarea class="form-control" rows="20">${json}</textarea>`);
-    });
-
-    $(document).on("click", ".search-result-item", function () {
-
-        const actionbarIndex = $(this).attr('data-actionbar-index');
-        const slotIndex = $(this).attr('data-slot-index');
-        const actionIndex = $(this).attr('data-action-index');
-        const itemId = $(this).attr('data-item-id');
-        const imageLink = $(this).find('.search-image').attr('src');
-
-        itemFetcher.fetchItemActionsById(itemId).then((actions) => {
-
-            if (actions.length == 1) {
-
-                profileManager.updateSlotAction(actionbarIndex, slotIndex, actions[0], actionIndex)
-                uiManager.setSlotAction(actionbarIndex, slotIndex, actionIndex, actions[0]);
-                uiManager.setSlotImage(actionbarIndex, slotIndex, imageLink);
-
-            } else {
-
-                modalProvider.showActionSelectionModal(actions, actionbarIndex, slotIndex, itemId, actionIndex, imageLink);
-
-            }
-        });
-    });
-
-    $(document).on("click", ".add-new-slot", function () {
-
-        var actionbarIndex = $(this).attr('data-actionbar-index');
-        var slotIndex = $(this).attr('data-slot-index');
-
-        modalProvider.showAddNewSlotModal(actionbarIndex, slotIndex);
-    });
-
-    $(document).on("click", "#add-actionbar-button", function () {   
-        var profileName = $('#profileDropdown').text();
-        profileManager.addEmptyActionbar(profileName);
-        loadActionbars();
-    });
-
-    $(document).on("click", ".delete-actionbar", function () {
-        var actionbarIndex = $(this).attr('data-actionbar-index');
-        var profileName = $('#profileDropdown').text();
-        
-        bootbox.confirm({
-            title: `Delete actionbar ${parseInt(actionbarIndex) + 1}`,
-            message: "Are you sure you want to delete this actionbar?",
-            buttons: {
-                confirm: {
-                    label: 'Yes',
-                    className: 'btn-danger'
-                },
-                cancel: {
-                    label: 'No',
-                    className: 'btn-secondary'
-                }
-            },
-            callback: function (result) {
-                if (result) {
-                    profileManager.deleteActionbar(profileName, actionbarIndex);
-                    loadActionbars();
-                }
-            }
-        });
     });
 
 });
